Rename classes and files for getting users with annotations, remove ordering
